Changes from code review - move logic from the component to the page controller and tidy up naming and flow
